requirements.txt
# CrewAI and core dependencies
crewai
langchain-openai
langchain-community

# Environment management
python-dotenv==1.0.0

# Required for CrewAI
litellm
pydantic>=2.0.0
pydantic_core

# API and web tools (updated for compatibility)
requests==2.31.0
aiohttp>=3.10.0
httpx

# Data processing
numpy==1.24.3

# Optional dependencies (uncomment as needed)
# pandas==2.0.3
# chromadb==0.4.22
# pinecone-client==2.2.4
# weaviate-client==3.25.3

# Testing (optional)
# pytest==7.4.3
# pytest-asyncio==0.21.1

# Development tools (optional)
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1